fix: Assign book categories from both views
Relocate state and API calls to App.js in order to be able to access the api update call from both - myReads view and searchResults view
Books can now be assigned to different categories and also be added to categories from the seachResults view as well as the library view
